+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.8] - 2020-03-07
+### Added
+
+- Add Logo Result, Name Result, and TVNetwork item (SRichter contributor)
+
+## [1.7.1] - 2020-03-07
+### Added
+
+- Add missing unit tests
+
 ## [1.7] - 2020-01-17
 ### Added